Fix NPE issue due to absent of keys and/or href elements.
As for topicref and keydef, the class is the same, hence there is a
chance a topicref with keyref attribute is present in the filtering
process.

Different from keydef, a topicref is valid with a keyref, without keys
and href attribute.

The process can't assume there is always a keys and href attribute for
only looking at the dita class "+ map/topicref mapgroup-d/keydef".